PMM-14577 Separe go-grpc-prometheus code into new file.
